Fix tests and add coverage
