+++ core/src/test/java/org/frankframework/soap/SoapWrapperTest.java
@@ -1,8 +1,6 @@
 package org.frankframework.soap;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertInstanceOf;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -15,25 +13,14 @@ import static org.mockito.Mockito.verify;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.math.BigInteger;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.KeyStore;
 import java.security.MessageDigest;
-import java.security.cert.Certificate;
-import java.security.cert.X509Certificate;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.Base64;
-import java.util.Date;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import javax.crypto.KeyGenerator;
-import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 import javax.xml.crypto.AlgorithmMethod;
 import javax.xml.crypto.KeySelector;
@@ -52,24 +39,15 @@ import javax.xml.transform.stream.StreamSource;
 import jakarta.xml.soap.MessageFactory;
 import jakarta.xml.soap.SOAPConstants;
 import jakarta.xml.soap.SOAPEnvelope;
-import jakarta.xml.soap.SOAPException;
 import jakarta.xml.soap.SOAPMessage;
 import jakarta.xml.soap.SOAPPart;
 
 import org.apache.wss4j.common.WSS4JConstants;
-import org.apache.wss4j.common.ext.WSSecurityException;
 import org.apache.wss4j.common.util.UsernameTokenUtil;
 import org.apache.wss4j.dom.WSConstants;
 import org.apache.wss4j.dom.WsuIdAllocator;
 import org.apache.xml.security.algorithms.JCEMapper;
-import org.bouncycastle.asn1.x500.X500Name;
-import org.bouncycastle.cert.jcajce.JcaX509CertificateConverter;
-import org.bouncycastle.cert.jcajce.JcaX509v3CertificateBuilder;
-import org.bouncycastle.operator.ContentSigner;
-import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -82,13 +60,9 @@ import lombok.extern.log4j.Log4j2;
 import org.frankframework.configuration.ConfigurationException;
 import org.frankframework.core.PipeLineSession;
 import org.frankframework.core.SenderException;
-import org.frankframework.encryption.KeystoreType;
-import org.frankframework.lifecycle.LoadBouncyCastleBean;
 import org.frankframework.stream.Message;
-import org.frankframework.stream.UrlMessage;
 import org.frankframework.testutil.MatchUtils;
 import org.frankframework.testutil.TestFileUtils;
-import org.frankframework.util.StreamUtil;
 import org.frankframework.util.StringUtil;
 import org.frankframework.util.UUIDUtil;
 import org.frankframework.util.XmlUtils;
@@ -318,134 +292,6 @@ public class SoapWrapperTest {
 		assertNotNull(file); // ensure we can find the file
 
 		assertTrue(verifySoapDigest(file.openStream()));
-	}
-
-	private KeyStore createDummyKeyStoreWithNullKeyPassword(String certificateName, String certificatePassword) throws Exception {
-		// Load BouncyCastle if not already set.
-		new LoadBouncyCastleBean().afterPropertiesSet();
-
-		KeyPairGenerator keyGen = KeyPairGenerator.getInstance("RSA");
-		keyGen.initialize(2048);
-		KeyPair keyPair = keyGen.generateKeyPair();
-
-		X500Name owner = new X500Name("CN=Test, OU=Test, O=Test, L=Test, C=US");
-		BigInteger serial = BigInteger.valueOf(System.currentTimeMillis());
-		Instant validFrom = Instant.now();
-		Instant validTo = validFrom.plus(365, ChronoUnit.DAYS);
-
-		JcaX509v3CertificateBuilder certBuilder = new JcaX509v3CertificateBuilder(
-				owner,
-				serial,
-				Date.from(validFrom),
-				Date.from(validTo),
-				owner,
-				keyPair.getPublic()
-		);
-		ContentSigner signer = new JcaContentSignerBuilder("SHA256withRSA").build(keyPair.getPrivate());
-		X509Certificate cert = new JcaX509CertificateConverter().setProvider("BC").getCertificate(certBuilder.build(signer));
-
-		KeyStore ks = KeyStore.getInstance(KeystoreType.PKCS12.name());
-		ks.load(null, "password".toCharArray());
-		ks.setKeyEntry(certificateName, keyPair.getPrivate(), certificatePassword.toCharArray(), new Certificate[] { cert } );
-		return ks;
-	}
-
-	@ParameterizedTest
-	@CsvSource({"true, true", "true, false", "false, true", "false, false"})
-	void validateEncryptedSoap1_1(boolean includeCertificateInMessage, boolean removeSecurityHeader) throws Exception {
-		URL file = TestFileUtils.getTestFileURL("/Soap/Encryption/SZeebraSoap.xml");
-		assertNotNull(file); // ensure we can find the file
-
-		String certificateName = "tralalal";
-		KeyStore keystore = createDummyKeyStoreWithNullKeyPassword(certificateName, "changeit");
-
-		KeyGenerator keyGen = KeyGenerator.getInstance("AES");
-		keyGen.init(256);
-		SecretKey secretKey = keyGen.generateKey();
-//		SecretKey secretKey = new SecretKeySpec(symmetricKey.getBytes(), "AES");
-
-		Message encrypted = SoapUtils.encryptMessage(new UrlMessage(file), keystore, certificateName, secretKey, includeCertificateInMessage);
-
-		String encryptedString = encrypted.asString()
-				.replaceAll("<xenc:CipherValue>.*?</xenc:CipherValue>", "<xenc:CipherValue>IGNORE-CIPHER-VALUE</xenc:CipherValue>")
-				.replaceAll("<wsu:Created>.*?</wsu:Created>", "<wsu:Created>IGNORE-CREATED</wsu:Created>")
-				.replaceAll("<wsu:Expires>.*?</wsu:Expires>", "<wsu:Expires>IGNORE-EXPIRES</wsu:Expires>")
-				.replaceAll("(Id=\")[^\"]*\"", "Id=\"id-here\"")
-				.replaceAll("(URI=\")[^\"]*\"", "URI=\"uri-here\"")
-				.replaceAll("(<wsse:BinarySecurityToken[^>]*>)[^<]*(</wsse:BinarySecurityToken>)", "$1IGNORE-BST$2")
-				.replaceAll("(<wsse:KeyIdentifier[^>]*>)[^<]*(</wsse:KeyIdentifier>)", "$1IGNORE-KI$2");
-
-		URL expectedFile = TestFileUtils.getTestFileURL("/Soap/Encryption/SZeebraSoap-encrypted-"+(includeCertificateInMessage ? "with" : "no")+"cert.xml");
-		assertNotNull(expectedFile); // ensure we can find the file
-		MatchUtils.assertXmlEquals(StreamUtil.resourceToString(expectedFile), encryptedString);
-
-		Message decrypted = SoapUtils.decryptMessage(encrypted, keystore, certificateName, "changeit", removeSecurityHeader);
-		// Ensure the decrypted result is the same as the initial document
-		MatchUtils.assertXmlEquals(removeSecurityHeader ? StreamUtil.resourceToString(file) : encrypted.asString(), decrypted.asString());
-	}
-
-	@Test
-	void validateEncryptedErrorSoap1_1() throws Exception {
-		URL file = TestFileUtils.getTestFileURL("/Soap/Encryption/SZeebraSoap.xml");
-		assertNotNull(file); // ensure we can find the file
-
-		String certificateName = "tralalal";
-		KeyStore keystore = createDummyKeyStoreWithNullKeyPassword(certificateName, "changeit");
-
-		KeyGenerator keyGen = KeyGenerator.getInstance("AES");
-		keyGen.init(256);
-		SecretKey secretKey = keyGen.generateKey();
-
-		Message encrypted = SoapUtils.encryptMessage(new UrlMessage(file), keystore, certificateName, secretKey, false);
-
-		WSSecurityException e1 = assertThrows(WSSecurityException.class, () -> SoapUtils.decryptMessage(encrypted, keystore, certificateName, "wrong-password", false));
-		assertEquals("unable to process security header", e1.getMessage());
-		assertNotNull(e1.getCause());
-		// Cause trace logs: "the private key for the supplied alias does not exist in the keystore"
-
-		WSSecurityException e2 = assertThrows(WSSecurityException.class, () -> SoapUtils.decryptMessage(encrypted, keystore, "wrong-cert", "changeit", false));
-		assertEquals("unable to process security header", e2.getMessage());
-		assertNotNull(e2.getCause());
-		// Cause trace logs: "the private key for the supplied alias does not exist in the keystore"
-
-		KeyStore differentStoreSameCertname = createDummyKeyStoreWithNullKeyPassword(certificateName, "changeit");
-		WSSecurityException e3 = assertThrows(WSSecurityException.class, () -> SoapUtils.decryptMessage(encrypted, differentStoreSameCertname, certificateName, "changeit", false));
-		assertEquals("unable to process security header", e3.getMessage());
-		assertNotNull(e3.getCause());
-		// Cause trace logs: "No certificates were found for decryption (KeyId)"
-
-		Message manipulatedMessage = new Message(encrypted.asString()
-				.replaceAll("<xenc:CipherValue>.*?</xenc:CipherValue>", "<xenc:CipherValue>IGNORE-CIPHER-VALUE</xenc:CipherValue>"));
-		WSSecurityException e4 = assertThrows(WSSecurityException.class, () -> SoapUtils.decryptMessage(manipulatedMessage, keystore, certificateName, "changeit", false));
-		assertEquals("unable to process security header", e4.getMessage());
-		assertNotNull(e3.getCause());
-		// Cause trace logs: "The signature or decryption was invalid"
-
-		// Swap the first 4 characters of the CipherValue
-		Message manipulatedMessage2 = new Message(encrypted.asString()
-				.replaceAll("<xenc:CipherValue>(.{4})(.{4})(.*)</xenc:CipherValue>", "<xenc:CipherValue>$2$1$3</xenc:CipherValue>"));
-		assertNotEquals(encrypted.asString(), manipulatedMessage2.asString());
-
-		Exception e5 = assertThrows(Exception.class, () -> SoapUtils.decryptMessage(manipulatedMessage2, keystore, certificateName, "changeit", false));
-		assertInstanceOf(WSSecurityException.class, e5, "expected a WSSecurityException but got: (%s): %s".formatted(e5.getClass(), e5.getMessage()));
-
-		Message resultMessage = SoapUtils.decryptMessage(new UrlMessage(file), keystore, certificateName, "changeit", false);
-		MatchUtils.assertXmlEquals(StreamUtil.resourceToString(file), resultMessage.asString());
-	}
-
-	// Reduce overhead when message is already of type SOAP
-	@Test
-	public void testMultipleToSoapCalls() throws SOAPException, IOException {
-		URL file = TestFileUtils.getTestFileURL("/Soap/Encryption/SZeebraSoap.xml");
-		assertNotNull(file); // ensure we can find the file
-
-		Message soapMessage = new UrlMessage(file);
-		// soapMessage.getContext().get(SoapContext.SOAP_VERSION_KEY);
-		final Document doc = SoapUtils.toSoapDocument(soapMessage);
-		Message soapDocumentMessage = new Message(doc);
-		final Document doc2 = SoapUtils.toSoapDocument(soapDocumentMessage);
-
-		assertEquals(doc, doc2, "expected both instances to be the same");
 	}
 
 	private Document toSoapMessage(InputStream is) throws Exception {
